requirements.txt
pycryptodome==3.4.3
